+++ cpp/arcae/table_factory.cc
@@ -86,6 +86,7 @@ Result<std::shared_ptr<NewTableProxy>> OpenTable(const std::string& filename,
 
 Result<std::shared_ptr<NewTableProxy>> DefaultMS(const std::string& name,
                                                  const std::string& subtable,
+                                                 std::size_t ninstances,
                                                  const std::string& json_table_desc,
                                                  const std::string& json_dminfo) {
   // Upper case subtable name
@@ -101,11 +102,13 @@ Result<std::shared_ptr<NewTableProxy>> DefaultMS(const std::string& name,
     modname.append(usubtable);
   }
 
-  ARROW_ASSIGN_OR_RAISE(
-      auto setup_new_table,
-      DefaultMSFactory(modname, usubtable, json_table_desc, json_dminfo));
+  auto MakeMS = [name = name, modname = modname, usubtable = usubtable,
+                 json_table_desc = json_table_desc,
+                 json_dminfo = json_dminfo]() -> Result<std::shared_ptr<TableProxy>> {
+    ARROW_ASSIGN_OR_RAISE(
+        auto setup_new_table,
+        DefaultMSFactory(modname, usubtable, json_table_desc, json_dminfo));
 
-  return NewTableProxy::Make([&]() -> Result<std::shared_ptr<TableProxy>> {
     // MAIN Measurement Set case
     if (usubtable.empty() || usubtable == kMain) {
       auto ms = MeasurementSet(setup_new_table);
@@ -172,7 +175,9 @@ Result<std::shared_ptr<NewTableProxy>> DefaultMS(const std::string& name,
     // Link the table against the Measurement Set
     ms.rwKeywordSet().defineTable(usubtable, subtable->table());
     return subtable;
-  });
+  };
+
+  return NewTableProxy::Make(std::move(MakeMS), ninstances);
 }
 
 // Execute a TAQL query on the supplied tables